style(src): Comments and sorting
